Fix exception if document isn't an Object
When matchDriveIdByCountry is called, occationally, the documents in
obi.availability aren't objects, but are strings, like 'Availability'.
This appears to be the old style, and all countries have more recent
entries that superseed this format. Except Moldova, which causes an
exception when creating cache/countries.json. This commit checks that
each document is an object before assigning drive/parent ids.